requirements.txt
aiogram==3.7.0
asyncpg==0.29.0
SQLAlchemy==2.0.30
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg[binary]==3.1.19
